--- processor/src/test/java/org/mapstruct/ap/test/bugs/_3809/ClassA.java
@@ -1,14 +1,0 @@
-/*
- * Copyright MapStruct Authors.
- *
- * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-package org.mapstruct.ap.test.bugs._3809;
-
-public class ClassA {
-    private NestedClassA param;
-
-    public NestedClassA getParam() {
-        return param;
-    }
-}
