--- processor/src/test/java/org/mapstruct/ap/test/bugs/_3809/NestedClassB.java
@@ -1,19 +1,0 @@
-/*
- * Copyright MapStruct Authors.
- *
- * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-package org.mapstruct.ap.test.bugs._3809;
-
-public class NestedClassB {
-    private String param1;
-
-    public String getParam1() {
-        return param1;
-    }
-
-    public void setParam1(String param1) {
-        this.param1 = param1;
-    }
-
-}
